+++ Packages/nudge/Runtime/Scripts/Comment.cs
@@ -15,7 +15,6 @@ namespace AID
     {
         private const string StartingBody = "...\n...";
 
-        [FormerlySerializedAs("hidden")]
         [SerializeField] private bool isHidden = false;
 
         [SerializeField] private bool isTask = false;
@@ -24,8 +23,8 @@ namespace AID
 
         [TextArea(5, 25)]
         [SerializeField] private string body = StartingBody;
-
-        [SerializeField] private UnityEngine.Object linkedObject;
+        [SerializeField] private UnityEngine.Object primaryLinkedObject;
+        [SerializeField] private UnityEngine.Object[] additionalLinkedObjects = System.Array.Empty<UnityEngine.Object>();
 #pragma warning restore CS0649
 
         [HideInInspector] [SerializeField] private string dateCreated;
@@ -42,10 +41,17 @@ namespace AID
 
         public void SetSelectedItems(UnityEngine.Object[] objs)
         {
-            linkedObject = objs[0];
+            primaryLinkedObject = objs[0];
+            additionalLinkedObjects = new Object[objs.Length - 1];
+            for (int i = 1; i < objs.Length; i++)
+            {
+                additionalLinkedObjects[i - 1] = objs[i];
+            }
         }
-
-        public UnityEngine.Object LinkedObject => linkedObject;
+		
+        public UnityEngine.Object PrimaryLinkedObject => primaryLinkedObject;
+        public System.Collections.ObjectModel.ReadOnlyCollection<UnityEngine.Object> AdditionalLinkedObjects => 
+            System.Array.AsReadOnly(additionalLinkedObjects);
         public bool Hidden => isHidden;
         public int Priority => priority;
         public bool IsTask => isTask;